repair some syntax

+++ src/controller/authController.go
@@ -151,7 +151,7 @@ func DisplayPendingAccount(c *gin.Context) {
 		"metadata": gin.H{
 			"total": total,
 			"page": page,
-			"last_page": math.Ceil(float64(int(total)/limit)),
+			"last_page": math.Ceil(float64(total)/float64(limit)),
 		},
 	})
 }
